dropped async functions and schema loading on run time
It was a bad design idea. I wanted to keep the browser bundle small and
tried to load missing schemas on the fly from a rest api. The amount of
async functions with multiple recursions was hard to debug and sometimes
it failed to load the data before running the tests. Now I preload the
schemas before running and save them to the localStorage of the browser.
So I have to load the 1,25MB only once and than it works out of the box.