started to implement string/object schema property fix
Some properties can be written in a short style like `"author":"John
Doe"`, but `author` is a property, that expects a range of type
`Organization` or `Person`, which are `Thing`s, that can have a `name`
property.

I'm a bit stuck here, because I'm not sure about the right way for
fixing it correctly. It feels like parsing in a non strict way and I
wasn't able to find a spec about transforming string values to typed
objects with a default property, yet.